+++ src/slang/ui/canvas/blueprint.ts
@@ -5,9 +5,9 @@ import {GenericPortModel, PortModel} from "../../core/abstract/port";
 import {SlangSubject} from "../../core/abstract/utils/events";
 import {BlueprintModel} from "../../core/models";
 import {TypeIdentifier} from "../../definitions/type";
-import {BlackBoxShape} from "../components/operator";
-import {ConnectionElement} from "../components/connection";
-import {BlueprintBox} from "../components/blueprint";
+import {BlackBoxShape} from "../elements/operator";
+import {ConnectionElement} from "../elements/connection";
+import {BlueprintBox} from "../elements/blueprint";
 import {Frame} from "../frame";
 import {Canvas, PaperViewArgs} from "./base";
 import {TargetableComponent, UserEvent } from "./user-events";
